Fix build on OSX
A few of the uses of abs causes build errors on OSX due to ambiguous
resolution. This is probably a compiler or libc++ bug since it is
trying to implicitly convert double to int without there being a
reason to.

Also, one use of std::auto_ptr caused an assignment initialisation
of a containing struct to fail to compile because it was trying to
implicitly copy construct auto_ptr. auto_ptr is generally a bad thing
anyway, so I've replaced it with std::shared_ptr here. It may be
possible to use std::unique_ptr here.