Fix: level strip width is wrong when you start open toonz (#334)
On the creation the docking panel of the level strip was assigned a minimum
width. This has no effect, but making the docking panel too big, because the
size of the underlying controls hasn't been calculated at this point.
Removing the line fixes the issue with the level strip size, without affecting
the docking window otherwise.