Tidy Topbar Style
Switch the bitmap used in #TopbarTab to a qlineargradient property. I
noticed it was already in the code but blocked out, was that due to
compile issues? I escaped it, so it should compile now if that was the
issue.

Also switch to inline comments when refering to less variables, so they
don't carry over to qss.