Color_0 chip redesign
This changes  the default color_0 chip to display a checkered pattern all the way to  the top.  This is for users who keep confusing the button for white.  Studio employee use probably knows what Color_0 stands for already. The change only affects the default chip.  All chips created afterwards with Color_0 will maintain the old style.

Co-Authored-By: turtletooth <jcbullock@gmail.com>